smoothed out all the kinks in time zone stuff
